refactor: merge rules in one file
